Rename eventConstantCheckingUtils to eventCheckingUtils
